# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "API documentation for the EduFlow pedagogical management solution.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EduFlow API Server"
        }
    ],
    "paths": {
...
        "/api/saved-courses": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "List all saved courses",
                "operationId": "e7bf82a07a441d12f16edbbe591da511",
                "responses": {
                    "200": {
                        "description": "List of saved courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "saved_courses": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/saved-courses/{course}": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add course to wishlist",
                "operationId": "55f31f720f58a4d72ed4d92a5a24321f",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Course saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "saved_course": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Course already saved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Remove course from wishlist",
                "operationId": "b622ad11dda0d9bbfb0dd6e7c1efb840",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course removed from saved courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saved course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token **_only_**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}